Refactor: Converter -> RoomDto에 정적 팩토리 메소드
