+++ src/main/java/com/group6/accommodation/domain/room/model/dto/RoomDto.java
@@ -1,7 +1,10 @@
 package com.group6.accommodation.domain.room.model.dto;
 
+import com.group6.accommodation.domain.room.model.entity.RoomEntity;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -37,4 +40,37 @@ public class RoomDto {
 	private String images;
 	private LocalDateTime checkInTime;
 	private LocalDateTime checkOutTime;
+
+	public static RoomDto toDto(RoomEntity room) {
+
+		return RoomDto.builder()
+			.roomId(room.getRoomId())
+			.accommodationId(room.getAccommodation().getId())
+			.title(room.getTitle())
+			.baseCount(room.getBaseCount())
+			.maxHeadCount(room.getMaxHeadCount())
+			.weekdaysFee(room.getWeekdaysFee())
+			.weekendsFee(room.getWeekendsFee())
+			.bath(room.isBath())
+			.hometheater(room.isHometheater())
+			.aircondition(room.isAircondition())
+			.tv(room.isTv())
+			.pc(room.isPc())
+			.cable(room.isCable())
+			.internet(room.isInternet())
+			.refrigerator(room.isRefrigerator())
+			.toiletries(room.isToiletries())
+			.sofa(room.isSofa())
+			.cook(room.isCook())
+			.table(room.isTable())
+			.hairdryer(room.isHairdryer())
+			.images(room.getImages())
+			.checkInTime(room.getCheckInTime())
+			.checkOutTime(room.getCheckOutTime())
+			.build();
+	}
+
+	public static List<RoomDto> toDtoList(List<RoomEntity> roomEntityList) {
+		return roomEntityList.stream().map(RoomDto::toDto).collect(Collectors.toList());
+	}
 }
